expecto/requirements.txt
biopython==1.79
cached-property==1.5.2
dataclasses==0.8
h5py==3.1.0
importlib-resources==5.4.0
natsort==8.2.0
numpy==1.19.5
pandas==0.22.0
Pillow==8.4.0
pkg_resources==0.0.0
pyfasta==0.5.2
python-dateutil==2.8.2
pytz==2023.3
scipy==1.5.4
six==1.16.0
tqdm==4.64.1
typing_extensions==4.1.1
xgboost==0.7.post4
zipp==3.6.0
